requirements.txt
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
jinja2
